+++ src/main/java/com/rae/creatingspace/server/entities/RoomShape.java
@@ -1,10 +1,7 @@
 package com.rae.creatingspace.server.entities;
 
 import net.minecraft.core.BlockPos;
-import net.minecraft.core.NonNullList;
 import net.minecraft.nbt.CompoundTag;
-import net.minecraft.network.FriendlyByteBuf;
-import net.minecraft.network.syncher.EntityDataSerializer;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.phys.AABB;
@@ -12,33 +9,7 @@ import net.minecraft.world.phys.AABB;
 import java.util.ArrayList;
 import java.util.List;
 
-import static net.minecraft.network.syncher.EntityDataSerializers.registerSerializer;
-
 public class RoomShape {
-
-    public static final EntityDataSerializer<RoomShape> SERIALIZER = new EntityDataSerializer.ForValueType<>() {
-        @Override
-        public void write(FriendlyByteBuf byteBuf, RoomShape roomShape) {
-            byteBuf.writeBoolean(roomShape.closed);
-            byteBuf.writeInt(roomShape.volume);
-            byteBuf.writeCollection(roomShape.listOfBox, ((byteBuf1, aabb) -> {
-                byteBuf1.writeBlockPos(new BlockPos(aabb.minX, aabb.minY, aabb.minZ));
-                byteBuf1.writeBlockPos(new BlockPos(aabb.maxX, aabb.maxY, aabb.maxZ));
-            }));
-        }
-
-        @Override
-        public RoomShape read(FriendlyByteBuf byteBuf) {
-            boolean closed = byteBuf.readBoolean();
-            int volume = byteBuf.readInt();
-            List<AABB> aabbList = byteBuf.readCollection(NonNullList::createWithCapacity, byteBuf1 -> new AABB(byteBuf1.readBlockPos(), byteBuf1.readBlockPos()));
-            return new RoomShape(aabbList, volume, closed);
-        }
-    };
-
-    static {
-        registerSerializer(SERIALIZER);
-    }
     ArrayList<AABB> listOfBox;
     boolean closed = true;
     float xRot;
